requirements.txt
requests>=2.31.0
PyJWT>=2.8.0
browser-cookie3>=0.19.1
beautifulsoup4>=4.12.0
lxml>=5.0.0
gspread>=6.0.0
google-auth>=2.0.0
python-dateutil>=2.9.0